File Browser + corrupt glbs implemented
Implemented a WHOLE lot of stuff. File Browser works now and allows to delete files. When this is done, the session and search browser don't refresh - they only refresh, when the deleted item is downloaded or reimported again, and then refresh all other items from the same source. To do this was completely bonkers and also required to add a txt for every glb that saves the according webLink.

IMPORTANT NOTE: we have to implement, that when leaving a session or entering one, lonesome pngs without GLBs are deleted. These are not deleted when a file is deleted, as they can be used within the session browser for objects, that are in the session, but have been recently deleted from the file browser.